Creado hook useMovieSearch en servicio
